refactor: :recycle: いいね関連のコードを切り出し
