+++ lib/feature/auth/application/auth_service.dart
@@ -5,8 +5,8 @@ import '../../../core/common_provider/is_loading_overlay_state.dart';
 import '../../../core/common_provider/toast_controller.dart';
 import '../../../util/logger.dart';
 import '../../definition/repository/fetch_definition_repository.dart';
-import '../../definition/repository/like_definition_repository.dart';
 import '../../definition/repository/write_definition_repository.dart';
+import '../../definition_like/repository/like_definition_repository.dart';
 import '../../user_config/application/user_config_state.dart';
 import '../../user_config/repository/device_info_repository.dart';
 import '../../user_config/repository/user_config_repository.dart';
