+++ lib/feature/definition/application/definition_state.dart
@@ -1,10 +1,10 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 import '../../auth/application/auth_state.dart';
+import '../../definition_like/repository/like_definition_repository.dart';
 import '../../user_profile/application/user_profile_state.dart';
 import '../domain/definition.dart';
 import '../repository/fetch_definition_repository.dart';
-import '../repository/like_definition_repository.dart';
 
 part 'definition_state.g.dart';
