+++ lib/feature/definition_like/repository/entity/like_document.freezed.dart
