+++ lib/page/definition_detail_page.dart
@@ -12,8 +12,8 @@ import '../core/common_widget/shimmer_widget.dart';
 import '../core/router/app_router.dart';
 import '../feature/auth/application/auth_state.dart';
 import '../feature/definition/application/definition_state.dart';
-import '../feature/definition/presentation/like_widget.dart';
 import '../feature/definition/presentation/self_definition_action_icon_button.dart';
+import '../feature/definition_like/presentation/like_widget.dart';
 import '../feature/user_follow/presentation/follow_or_unfollow_button.dart';
 import '../util/extension/date_time_extension.dart';
 import '../util/logger.dart';
